requirements.txt
# InstAntiBot requirements...
python-dotenv
instagrapi
# igramscraper=0.3.5 requirements...
wheel
Pillow
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.10
lxml
parse==1.18.0
pyee==7.0.4
pyppeteer==0.2.3
pyquery==1.4.3
python-slugify==3.0.2
requests==2.25.1
requests-html==0.10.0
six==1.15.0
soupsieve==2.1
text-unidecode==1.2
tqdm==4.55.1
urllib3==1.26.2
w3lib==1.22.0
websockets==8.1
# Install requirements...
.\igramscrapersrc\